First Update
Created Data Handling

The player is assigned a plot on join.

If its the first time the player has joined, open the FirstJoinWidget.

The player is shown an arial view of the plot while the widget is displayed.

When the user clicks on the "Open Restaurant" button, he is teleported to the plot.