Updated Warp, updated game flow
Updated Warp to latest version

Removed FirstJoinWidget

StartGameplay is called when all client modules are loaded

All relevant profile data is loaded to client including the PlotNumber. The structure of the data has been modified to make access to it easier to access.

The user is shown their plot from an arial view and shown the "open restaurant" GUI.